refactor(#487): deprecate old AbstractAnnotation component
