+++ annotation-interface/pages/annotations/medications/[id].tsx
@@ -12,7 +12,7 @@ import {
     serverEndpointMeds,
     ServerMedication,
 } from '../../../common/server-types';
-import { BackToAnnotations } from '../../../components/annotations/AbstractAnnotation';
+import { BackToAnnotations } from '../../../components/annotations/AbstractAnnotationOld';
 import { MedAnnotation } from '../../../components/annotations/BrickAnnotations';
 import PageHeading from '../../../components/common/PageHeading';
 import dbConnect from '../../../database/helpers/connect';
